spellcheck, update electron-storage, update electron
