+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ChunChunMaru",
-  "version": "1.1.3",
+  "version": "2.0.0",
   "description": "get your own dashboard",
   "main": "./script/electron.js",
   "scripts": {
@@ -18,12 +18,13 @@
   "author": "Meystin",
   "license": "CC0-1.0",
   "devDependencies": {
-    "electron-builder": "20.28.4",
-    "electron-packager": "12.1.1"
+    "electron": "^9.1.1",
+    "electron-builder": "22.11.7"
   },
   "dependencies": {
-    "ajv": "^6.10.2",
-    "electron-json-storage": "^4.1.8"
+    "ajv": "^6.12.6",
+    "electron-packager": "15.3.0",
+    "electron-storage": "1.0.7"
   },
   "build": {
     "appId": "my.ChunChunMaru"
