--- script/animeview.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>ChunChunMaru WishList</title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link href="bootstrap.css" rel="stylesheet">
-    </head>
-    <body>
-        <div id="newAnime" style="display:none">
-            <input type='button' onclick="getUrl()" value="Sauvegarde"/>
-            <input type='button' onclick="getStateAnime()" value="Anime"/>
-        <webview id='search' src='https://www.nautiljon.com/' style="width:100%"></webview></div>
-        <div id="wishList">
-            <input type='button' onclick="ajout()" value="Ajout"/>
-        </div>
-        <input type="button" value="delete" onclick="deletedb()"/> 
-        <script src="anime.js"></script>
-    </body>
-    
-</html>
