+++ script/chatParam/chatview.html
@@ -1,0 +1,18 @@
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <title>ChunChunMaru Dashboard</title>
+    </head>
+    <body>
+        <webview id="1">
+        </webview><webview id="2">
+        </webview><webview id="3">
+        </webview><webview id="4"></webview>
+        <img name="1" src="../../img/1.jpg" onmouseup="getEvent(event, this)" height="50" width="50"/>
+        <img name="2" src="../../img/2.jpg" onmouseup="getEvent(event, this)" height="50" width="50"/>
+        <img name="3" src="../../img/3.jpg" onmouseup="getEvent(event, this)" height="50" width="50"/>
+        <img name="4" src="../../img/4.jpg" onmouseup="getEvent(event, this)" height="50" width="50"/>
+        <img src="../../img/overview.jpg" onclick="overview()" height="50" width="50"/>
+        <script src="chat.js"></script>
+</body>
+</html>
