--- script/urlAnime.json
@@ -1,1 +1,0 @@
-["{\"titre\":\"emiya-san chi no kyou no gohan\",\"url\":\"https://www.nautiljon.com/animes/emiya-san+chi+no+kyou+no+gohan.html\"}","{\"titre\":\"cupid-s chocolates 2nd season\",\"url\":\"https://www.nautiljon.com/animes/cupid-s+chocolates+2nd+season.html\"}"]
